Allow use of CloudFront with a pre-set path
When using CloudFront manually with the Origin set to the exact stage
URL, the WSGI application thinks that it is mounted at `/${stage}`
rather than `/`. This causes it to generate and return invalid URLs to
the user.

Example:
A CF distribution for `example.com` pointing at
`0deadbeef.execute-api.eu-west-1.amazonaws.com/production` should think
that `/foo` is mounted at `https://example.com/foo` not
`https://example.com/production/foo`

We need to fix this behaviour due to the assumption that all requests
with `amazonaws.com` in the Host header will contain the stage name in
the path. This is not true for CloudFront as it will send the `Host`
header of the origin, triggering our `if` clause.